Feat
Security is here!
Add security validation for the endpoints, and the users must logging
with its twitter account. Most of the magic is done by spring.
Certain things can be improved (and will later on), but this first
version will do for now.
Add a table creation script, as the security feature needs an SQL table
(UserConnection), to store the users connections.